Move cli-version.ts to /codeql-cli/cli-version.ts
