+++ extensions/ql-vscode/src/distribution.ts
@@ -13,7 +13,7 @@ import {
   showAndLogWarningMessage,
 } from "./helpers";
 import { extLogger } from "./common";
-import { getCodeQlCliVersion } from "./cli-version";
+import { getCodeQlCliVersion } from "./codeql-cli/cli-version";
 import { ProgressCallback, reportStreamProgress } from "./progress";
 import {
   codeQlLauncherName,
